feat: add basic email templates to platform
