+++ examples/minimal/package.json
@@ -32,10 +32,10 @@
   "dependencies": {
     "@babel/runtime": "^7.9.6",
     "@unchainedshop/controlpanel": "^0.48.1",
-    "apollo-server-express": "^2.13.0",
+    "apollo-server-express": "^2.13.1",
     "bcrypt": "^4.0.1",
-    "braintree": "^2.22.0",
-    "emoji-flags": "^1.2.0",
+    "braintree": "^2.23.0",
+    "emoji-flags": "^1.3.0",
     "faker": "^4.1.0",
     "graphql": "^14.6.0",
     "i18n-iso-countries": "^5.3.0",
@@ -52,7 +52,7 @@
     "open": "^7.0.3",
     "ramda": "^0.27.0",
     "simpl-schema": "^1.7.1",
-    "stripe": "^8.50.0",
+    "stripe": "^8.52.0",
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
