--- examples/minimal/templates/index.js
@@ -1,4 +1,0 @@
-import './order-confirmation';
-import './forward-delivery';
-import './quotation-status';
-import './account-action';
